refactor: move linker user role tasks
