+++ handlers/linker/linkerAdminUserLevelsPage.go
@@ -4,19 +4,14 @@ import (
 	"context"
 	"database/sql"
 	"errors"
-	"fmt"
 	"github.com/arran4/goa4web/core/consts"
 	"log"
 	"net/http"
-	"strconv"
 	"strings"
 
 	"github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/handlers"
 	"github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/eventbus"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/internal/tasks"
 )
 
 func AdminUserRolesPage(w http.ResponseWriter, r *http.Request) {
@@ -78,85 +73,6 @@ func AdminUserRolesPage(w http.ResponseWriter, r *http.Request) {
 	handlers.TemplateHandler(w, r, "adminUserRolesPage.gohtml", data)
 }
 
-type userAllowTask struct{ tasks.TaskString }
-
-var UserAllowTask = &userAllowTask{TaskString: TaskUserAllow}
-
-var _ notif.TargetUsersNotificationProvider = (*userAllowTask)(nil)
-var _ tasks.Task = (*userAllowTask)(nil)
-
-func (userAllowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	usernames := r.PostFormValue("usernames")
-	role := r.PostFormValue("role")
-	fields := strings.FieldsFunc(usernames, func(r rune) bool {
-		return r == ',' || r == '\n' || r == '\r' || r == '\t' || r == ' '
-	})
-	for _, username := range fields {
-		if username == "" {
-			continue
-		}
-		u, err := queries.GetUserByUsername(r.Context(), sql.NullString{Valid: true, String: username})
-		if err != nil {
-			log.Printf("GetUserByUsername Error: %s", err)
-			continue
-		}
-		if err := queries.CreateUserRole(r.Context(), db.CreateUserRoleParams{
-			UsersIdusers: u.Idusers,
-			Name:         role,
-		}); err != nil {
-			log.Printf("permissionUserAllow Error: %s", err)
-		} else if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-			if evt := cd.Event(); evt != nil {
-				if evt.Data == nil {
-					evt.Data = map[string]any{}
-				}
-				evt.Data["targetUserID"] = u.Idusers
-				evt.Data["Username"] = u.Username.String
-				evt.Data["Role"] = role
-			}
-		}
-	}
-	return nil
-}
-
-type userDisallowTask struct{ tasks.TaskString }
-
-var UserDisallowTask = &userDisallowTask{TaskString: TaskUserDisallow}
-
-var _ notif.TargetUsersNotificationProvider = (*userDisallowTask)(nil)
-var _ tasks.Task = (*userDisallowTask)(nil)
-
-func (userDisallowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	r.ParseForm()
-	ids := r.Form["permids"]
-	if len(ids) == 0 {
-		if id := r.PostFormValue("permid"); id != "" {
-			ids = append(ids, id)
-		}
-	}
-	for _, idStr := range ids {
-		permid, _ := strconv.Atoi(idStr)
-		infoID, username, role, err2 := roleInfoByPermID(r.Context(), queries, int32(permid))
-		if err := queries.DeleteUserRole(r.Context(), int32(permid)); err != nil {
-			log.Printf("permissionUserDisallow Error: %s", err)
-		} else if err2 == nil {
-			if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-				if evt := cd.Event(); evt != nil {
-					if evt.Data == nil {
-						evt.Data = map[string]any{}
-					}
-					evt.Data["targetUserID"] = infoID
-					evt.Data["Username"] = username
-					evt.Data["Role"] = role
-				}
-			}
-		}
-	}
-	return nil
-}
-
 func roleInfoByPermID(ctx context.Context, q *db.Queries, id int32) (int32, string, string, error) {
 	rows, err := q.GetPermissionsWithUsers(ctx, db.GetPermissionsWithUsersParams{Username: sql.NullString{}})
 	if err != nil {
@@ -168,42 +84,4 @@ func roleInfoByPermID(ctx context.Context, q *db.Queries, id int32) (int32, stri
 		}
 	}
 	return 0, "", "", sql.ErrNoRows
-}
-
-func (userAllowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
-	if id, ok := evt.Data["targetUserID"].(int32); ok {
-		return []int32{id}, nil
-	}
-	if id, ok := evt.Data["targetUserID"].(int); ok {
-		return []int32{int32(id)}, nil
-	}
-	return nil, fmt.Errorf("target user id not provided")
-}
-
-func (userAllowTask) TargetEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("setUserRoleEmail")
-}
-
-func (userAllowTask) TargetInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("set_user_role")
-	return &v
-}
-
-func (userDisallowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
-	if id, ok := evt.Data["targetUserID"].(int32); ok {
-		return []int32{id}, nil
-	}
-	if id, ok := evt.Data["targetUserID"].(int); ok {
-		return []int32{int32(id)}, nil
-	}
-	return nil, fmt.Errorf("target user id not provided")
-}
-
-func (userDisallowTask) TargetEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("deleteUserRoleEmail")
-}
-
-func (userDisallowTask) TargetInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("delete_user_role")
-	return &v
 }
