+++ handlers/linker/user_disallow_task.go
@@ -1,0 +1,72 @@
+package linker
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"strconv"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/internal/eventbus"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// userDisallowTask removes a user's role.
+type userDisallowTask struct{ tasks.TaskString }
+
+// UserDisallowTask is the exported instance used when registering routes.
+var UserDisallowTask = &userDisallowTask{TaskString: TaskUserDisallow}
+
+var _ tasks.Task = (*userDisallowTask)(nil)
+var _ notif.TargetUsersNotificationProvider = (*userDisallowTask)(nil)
+
+func (userDisallowTask) Action(w http.ResponseWriter, r *http.Request) any {
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	r.ParseForm()
+	ids := r.Form["permids"]
+	if len(ids) == 0 {
+		if id := r.PostFormValue("permid"); id != "" {
+			ids = append(ids, id)
+		}
+	}
+	for _, idStr := range ids {
+		permid, _ := strconv.Atoi(idStr)
+		infoID, username, role, err2 := roleInfoByPermID(r.Context(), queries, int32(permid))
+		if err := queries.DeleteUserRole(r.Context(), int32(permid)); err != nil {
+			log.Printf("permissionUserDisallow Error: %s", err)
+		} else if err2 == nil {
+			if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
+				if evt := cd.Event(); evt != nil {
+					if evt.Data == nil {
+						evt.Data = map[string]any{}
+					}
+					evt.Data["targetUserID"] = infoID
+					evt.Data["Username"] = username
+					evt.Data["Role"] = role
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func (userDisallowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
+	if id, ok := evt.Data["targetUserID"].(int32); ok {
+		return []int32{id}, nil
+	}
+	if id, ok := evt.Data["targetUserID"].(int); ok {
+		return []int32{int32(id)}, nil
+	}
+	return nil, fmt.Errorf("target user id not provided")
+}
+
+func (userDisallowTask) TargetEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("deleteUserRoleEmail")
+}
+
+func (userDisallowTask) TargetInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("delete_user_role")
+	return &v
+}
